fix: lint cache issues
