+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -150,7 +150,6 @@ import {
 import { Composer, RouterConfigUploadError } from '../composition/composer.js';
 import { buildSchema, composeSubgraphs } from '../composition/composition.js';
 import { getDiffBetweenGraphs } from '../composition/schemaCheck.js';
-import { schemaLintCheck } from '../composition/schemaLint.js';
 import { audiences, nowInSeconds, signJwtHS256 } from '../crypto/jwt.js';
 import { AuthenticationError, PublicError } from '../errors/errors.js';
 import { OpenAIGraphql } from '../openai-graphql/index.js';
@@ -210,6 +209,7 @@ import {
 } from '../util.js';
 import { FederatedGraphSchemaUpdate, OrganizationWebhookService } from '../webhooks/OrganizationWebhookService.js';
 import { apiKeyPermissions } from '../constants.js';
+import SchemaLinter from '../services/SchemaLinter.js';
 
 export default function (opts: RouterOptions): Partial<ServiceImpl<typeof PlatformService>> {
   return {
@@ -1697,7 +1697,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         if (namespace.enableLinting && newSchemaSDL !== '') {
           const lintConfigs = await schemaLintRepo.getNamespaceLintConfig(namespace.id);
           if (lintConfigs.length > 0) {
-            lintIssues = await schemaLintCheck({
+            const schemaLinter = new SchemaLinter();
+            lintIssues = await schemaLinter.schemaLintCheck({
               schema: newSchemaSDL,
               rulesInput: lintConfigs,
             });
